ml-services/api-gateway/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.32.3
python-dotenv==1.0.0
gunicorn==21.2.0
pyjwt==2.8.0